# pb/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "hellow.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Greeter"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/say/hello": {
      "post": {
        "summary": "Sends a greeting",
        "operationId": "Greeter_SayHello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/grpc_demoHelloReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "The request message containing the user's name.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/grpc_demoHelloRequest"
            }
          }
        ],
        "tags": [
          "Greeter"
        ]
      }
    }
  },
  "definitions": {
    "grpc_demoHelloReply": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "The response message containing the greetings"
    },
    "grpc_demoHelloRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "hobbies": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      },
      "description": "The request message containing the user's name."
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}